requirements.txt
Flask-RESTful==0.3.7
flask-sqlalchemy==2.4.0
flask==1.0.2
flask-script==2.0.6
flask-migrate==2.5.2
pymysql==0.9.3
requests==2.22.0
python-dotenv==0.10.1
